btl/requirements.txt
dotenv
openai
google-generativeai